requirements.txt
aiohttp==3.7.3
PyYAML==5.4.1
